Added analyze-jumpstats-optimized.py
This is a recode of the first version.
This one doesn't use a seperate loop for each analytic, but instead,
uses 1 loop and calculates all the analytics in that one.
This should make it run a bit faster.

New version doesn't support command-line arguments *yet*, but I'll
add those soon.